# swagger.json
{"openapi":"3.0.0","paths":{"/auth/test":{"get":{"operationId":"AuthController_test","parameters":[],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"User login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in."}},"tags":["Authentication"]}},"/auth/register":{"post":{"operationId":"AuthController_register","summary":"User registration","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered."}},"tags":["Authentication"]}},"/auth/logout":{"get":{"operationId":"AuthController_logout","summary":"User logout","parameters":[],"responses":{"200":{"description":"User successfully logged out."}},"tags":["Authentication"]}},"/auth/generate/tokens":{"get":{"operationId":"AuthController_generateTokens","parameters":[],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","summary":"Request password reset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"Password reset request sent successfully."}},"tags":["Authentication"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","summary":"Reset user password","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"User password reset successfully."}},"tags":["Authentication"]}},"/auth/verify-email":{"get":{"operationId":"AuthController_verifyEmail","summary":"Verify user email","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"User email verified successfully."}},"tags":["Authentication"]}},"/auth/user-role":{"get":{"operationId":"AuthController_userRole","summary":"Get user role","parameters":[],"responses":{"200":{"description":"Returns the user role."}},"tags":["Authentication"]}},"/auth/profile":{"get":{"operationId":"AuthController_profile","summary":"Get user profile","parameters":[],"responses":{"200":{"description":"Returns the user profile."}},"tags":["Authentication"]}},"/auth/send-verification-email":{"post":{"operationId":"AuthController_sendVerificationEmail","summary":"Send email verification link","parameters":[],"responses":{"200":{"description":"Verification email sent successfully."}},"tags":["Authentication"]}},"/users":{"post":{"operationId":"UserController_create","summary":"Create a user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"The user has been successfully created."}},"tags":["Users"]},"get":{"operationId":"UserController_findAll","summary":"Find users with pagination","parameters":[{"name":"search","required":false,"in":"query","description":"Search Word","schema":{"example":"John","type":"string"}},{"required":false,"description":"Search Word","name":"search","in":"query","schema":{"example":"John","type":"string"}}],"responses":{"200":{"description":"Returns a list of users with pagination."}},"tags":["Users"]}},"/users/me":{"get":{"operationId":"UserController_getLoggedinUserDetails","summary":"Get logged-in user details","parameters":[],"responses":{"200":{"description":"Returns details of the logged-in user."}},"tags":["Users"]},"delete":{"operationId":"UserController_deleteLoggedinUserDetails","summary":"Delete logged-in user details","parameters":[],"responses":{"204":{"description":"User details have been successfully deleted."}},"tags":["Users"]}},"/users/me/images/upload":{"post":{"operationId":"UserController_uploadLoggedinUserImage","summary":"Upload an image for the logged-in user","parameters":[],"responses":{"200":{"description":"Image has been successfully uploaded."}},"tags":["Users"]}},"/users/{id}":{"put":{"operationId":"UserController_updateById","summary":"Update a user by ID","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"get":{"operationId":"UserController_findById","summary":"Find a user by ID","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UserController_deleteById","summary":"Delete a user by ID","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/{id}/images/upload":{"post":{"operationId":"UserController_upload","summary":"Upload an image for a user by ID","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Image has been successfully uploaded."}},"tags":["Users"]}},"/cards":{"post":{"operationId":"CardController_create","summary":"Create a new Card","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardDto"}}}},"responses":{"201":{"description":"The Card has been successfully created."}},"tags":["Card"],"security":[{"bearer":[]}]}},"/cards/review":{"patch":{"operationId":"CardController_reviewCard","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardReviewDto"}}}},"responses":{"200":{"description":""}},"tags":["Card"]}},"/cards/next-card":{"get":{"operationId":"CardController_nextCard","summary":"Next Card","parameters":[],"responses":{"200":{"description":"Successfully found Card"}},"tags":["Card"]}},"/cards/{id}":{"get":{"operationId":"CardController_findById","summary":"Find Card by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully found the Card"},"404":{"description":"Card not found"}},"tags":["Card"]},"put":{"operationId":"CardController_updateById","summary":"Update Card by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardDto"}}}},"responses":{"200":{"description":"Successfully updated the Card"},"404":{"description":"Card not found"}},"tags":["Card"],"security":[{"bearer":[]}]},"delete":{"operationId":"CardController_deleteById","summary":"Delete Card by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully deleted the Card"},"404":{"description":"Card not found"}},"tags":["Card"],"security":[{"bearer":[]}]}},"/decks":{"post":{"operationId":"DeckController_create","summary":"Create a new Deck","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeckDto"}}}},"responses":{"201":{"description":"The Deck has been successfully created."}},"tags":["Deck"],"security":[{"bearer":[]}]},"get":{"operationId":"DeckController_findAll","summary":"Find all Deck","parameters":[],"responses":{"200":{"description":"Successfully found Deck"}},"tags":["Deck"]}},"/decks/{id}":{"get":{"operationId":"DeckController_findById","summary":"Find Deck by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully found the Deck"},"404":{"description":"Deck not found"}},"tags":["Deck"]},"put":{"operationId":"DeckController_updateById","summary":"Update Deck by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDeckDto"}}}},"responses":{"200":{"description":"Successfully updated the Deck"},"404":{"description":"Deck not found"}},"tags":["Deck"],"security":[{"bearer":[]}]},"delete":{"operationId":"DeckController_deleteById","summary":"Delete Deck by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully deleted the Deck"},"404":{"description":"Deck not found"}},"tags":["Deck"],"security":[{"bearer":[]}]}}},"info":{"title":"API","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"localhost:8080"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"User email","example":"user@example.com","uniqueItems":true,"format":"email","minLength":5,"maxLength":60,"nullable":false},"password":{"type":"string","description":"User password","example":"123456","minLength":6,"maxLength":60,"nullable":false}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name","example":"John","maxLength":50,"minLength":2},"lastName":{"type":"string","description":"Last name","example":"Doe","maxLength":50,"minLength":2},"role":{"type":"string","description":"Role","enum":["User","Admin","All"],"example":"User"},"email":{"type":"string","description":"Email","example":"be@example.com","maxLength":50,"minLength":5,"format":"email"},"password":{"type":"string","description":"Password","example":"123456","maxLength":50,"minLength":6}},"required":["firstName","lastName","role","email","password"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"Email","example":"be@example.com","maxLength":50,"minLength":5,"format":"email"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"password":{"type":"string","description":"New Passowrd","example":"123456"},"passwordConfirmation":{"type":"string","description":"Confirm Password","example":"123456"}},"required":["password","passwordConfirmation"]},"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"User first name","example":"John"},"lastName":{"type":"string","description":"User last name","example":"Doe"},"email":{"type":"string","description":"User email","example":"be@example.com"},"password":{"type":"string","description":"User password","example":"123456"},"role":{"type":"string","description":"User role","enum":["User","Admin","All"],"example":"User"}},"required":["firstName","lastName","email","password","role"]},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"User first name","example":"John"},"lastName":{"type":"string","description":"User last name","example":"Doe"},"email":{"type":"string","description":"User email","example":"be@example.com"},"password":{"type":"string","description":"User password","example":"123456"},"role":{"type":"string","description":"User role","enum":["User","Admin","All"],"example":"User"}}},"CreateCardDto":{"type":"object","properties":{}},"CardReviewDto":{"type":"object","properties":{}},"UpdateCardDto":{"type":"object","properties":{}},"CreateDeckDto":{"type":"object","properties":{}},"UpdateDeckDto":{"type":"object","properties":{}}}}}